update to user authorized delete

+++ WishList/Controllers/ItemController.cs
@@ -44,9 +44,17 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
-            _context.SaveChanges();
+            if (item.User.Id != user.Id)
+            {
+                return Unauthorized();                
+            }
+            else
+            {
+                _context.Items.Remove(item);
+                _context.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
     }
